docs(json): 更新 JSON 模块文档注释并调整解析选项
- 为 Json 类添加详细的功能说明和使用示例
- 更新 JsonBeanMapper、JsonDecoder、JsonEncoder 接口的文档描述
- 为 JsonException、JsonFeature 枚举添加完整的功能说明
- 重构 JsonOptions 配置类，新增严格模式和宽松模式预设常量
- 优化 JsonParser 解析器的内部实现和错误处理机制
- 为 JsonPath 路径解析工具添加 DLZ 路径语法支持
- 增强 JsonText 快速形态判断工具的功能说明
- 完善 JsonTypes 泛型类型描述工具的文档注释

+++ dlz-kit/src/main/java/com/dlz/kit/json/core/JsonWriter.java
@@ -6,20 +6,42 @@ import java.util.IdentityHashMap;
 import java.util.Iterator;
 import java.util.Map;
 
+/**
+ * JSON 序列化器（语法层，包内私有）。
+ *
+ * <p>把 Java 对象递归写成 JSON 文本：Map→对象、Iterable/数组→数组、字符串→带引号字符串、
+ * 数字→数字、布尔→布尔、null→null、枚举→其 name()。使用 {@link IdentityHashMap} 检测
+ * 循环引用，避免无限递归。行为受 {@link JsonOptions} 控制（null 输出、美化、非 ASCII 转义等）。</p>
+ */
 final class JsonWriter {
+    /** 序列化选项（特性与上限）。 */
     private final JsonOptions options;
+    /** 输出缓冲区。 */
     private final StringBuilder output = new StringBuilder();
+    /** 正在写入的对象集合，用于检测循环引用（按对象身份去重）。 */
     private final IdentityHashMap<Object, Boolean> visiting = new IdentityHashMap<Object, Boolean>();
 
     JsonWriter(JsonOptions options) {
         this.options = options;
     }
 
+    /**
+     * 入口：把值序列化为 JSON 文本字符串。
+     *
+     * @param value 任意 Java 对象
+     * @return JSON 文本
+     */
     String write(Object value) {
         writeValue(value, 0);
         return output.toString();
     }
 
+    /**
+     * 按类型分派，写入单个值。
+     *
+     * @param value 待写入的值
+     * @param depth 当前嵌套深度
+     */
     private void writeValue(Object value, int depth) {
         checkDepth(depth);
         if (value == null) {
@@ -41,6 +63,12 @@ final class JsonWriter {
         }
     }
 
+    /**
+     * 写入 JSON 对象 {@code {...}}，键保持顺序；未开启 WRITE_NULLS 时跳过 null 值字段。
+     *
+     * @param value 源映射
+     * @param depth 当前嵌套深度
+     */
     private void writeObject(Map<?, ?> value, int depth) {
         checkContainerSize(value.size());
         enter(value);
@@ -72,6 +100,13 @@ final class JsonWriter {
         }
     }
 
+    /**
+     * 写入 JSON 数组 {@code [...]}（来源为 Iterable）。
+     *
+     * @param iterator 元素迭代器
+     * @param identity 用于循环检测的原始集合对象
+     * @param depth    当前嵌套深度
+     */
     private void writeArray(Iterator<?> iterator, Object identity, int depth) {
         enter(identity);
         try {
@@ -96,6 +131,12 @@ final class JsonWriter {
         }
     }
 
+    /**
+     * 写入 JSON 数组（来源为 Java 原生数组，使用反射逐个取值）。
+     *
+     * @param value 源数组
+     * @param depth 当前嵌套深度
+     */
     private void writeJavaArray(Object value, int depth) {
         enter(value);
         try {
@@ -119,10 +160,10 @@ final class JsonWriter {
     }
 
     private void writeNumber(Number value) {
-        if (value instanceof Double && !Double.isFinite(((Double) value).doubleValue())) {
+        if (value instanceof Double && !Double.isFinite((value).doubleValue())) {
             throw new JsonException("Non-finite double is not valid JSON");
         }
-        if (value instanceof Float && !Float.isFinite(((Float) value).floatValue())) {
+        if (value instanceof Float && !Float.isFinite((value).floatValue())) {
             throw new JsonException("Non-finite float is not valid JSON");
         }
         if (value instanceof BigDecimal) {
@@ -132,6 +173,11 @@ final class JsonWriter {
         }
     }
 
+    /**
+     * 写入带双引号的字符串，转义控制字符与引号；开启 ESCAPE_NON_ASCII 时转义非 ASCII。
+     *
+     * @param value 源字符串
+     */
     private void writeString(String value) {
         if (value.length() > options.getMaxStringLength()) {
             throw new JsonException("String exceeds maxStringLength " + options.getMaxStringLength());
@@ -158,6 +204,7 @@ final class JsonWriter {
         output.append('"');
     }
 
+    /** 把一个字符写成 {@code \\uXXXX} 形式的转义序列（不足 4 位前补 0）。 */
     private void appendUnicode(char value) {
         output.append("\\u");
         String hex = Integer.toHexString(value);
@@ -167,6 +214,7 @@ final class JsonWriter {
         output.append(hex);
     }
 
+    /** 开启 WRITE_PRETTY 时换行并按层级缩进（每层两个空格）。 */
     private void newlineAndIndent(int depth) {
         if (!enabled(JsonFeature.WRITE_PRETTY)) {
             return;
@@ -177,28 +225,33 @@ final class JsonWriter {
         }
     }
 
+    /** 进入一个容器对象：若已处于访问中说明存在循环引用，立即报错。 */
     private void enter(Object value) {
         if (visiting.put(value, Boolean.TRUE) != null) {
             throw new JsonException("Circular reference detected while writing JSON");
         }
     }
 
+    /** 离开容器对象：从访问集合中移除，配合 {@link #enter(Object)} 使用。 */
     private void leave(Object value) {
         visiting.remove(value);
     }
 
+    /** 校验嵌套深度是否超过上限。 */
     private void checkDepth(int depth) {
         if (depth > options.getMaxDepth()) {
             throw new JsonException("JSON nesting depth exceeds " + options.getMaxDepth());
         }
     }
 
+    /** 校验容器元素个数是否超过上限。 */
     private void checkContainerSize(int size) {
         if (size > options.getMaxContainerSize()) {
             throw new JsonException("JSON container size exceeds " + options.getMaxContainerSize());
         }
     }
 
+    /** 查询某序列化特性是否开启。 */
     private boolean enabled(JsonFeature feature) {
         return options.isEnabled(feature);
     }
